Fix quote-text backdrop-filter, fix overflow issue
Backdrop-filter was only limited to the quote-sub-section, which
resulted in it not covering the full width of the background image.
This is because the section was set to a max-width of 960px.
This hard limit also caused this section to overflow when reducing the
window width, resulting in a white blank area appearing to the right
of the content.

Remove max-width from quote-sub-section then readjust the quote-text
max-width to match more with the project goal picture.