Fix overflow issue from .quote-source-text
Set display to inline-block.

Cause:
Since the text is a block level element, the width equals the full
width of its container.
When adjusting its position, the empty space from its width causes
blank space to be present on the right.
In other words, the element is overflowing past its parent container.